pyproject.toml
[tool.poetry]
name = "halodedupe"
version = "0.1.0"
description = ""
authors = ["zolizoli <zoltan.varju@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.2"
floret = "^0.10.3"
numpy = "^1.24.3"
umap-learn = "^0.5.3"
scikit-learn = "^1.2.2"
altair = "^5.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
vega-datasets = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
